+++ app/src/main/java/com/example/mytodo/data/model/Plan.java
@@ -1,16 +1,20 @@
-package com.example.mytodo;
+package com.example.mytodo.data.model;
 
 import java.util.Calendar;
 
-public class Task {
+public class Plan {
   private String title;
   private String description;
   private Calendar CalendarStart;
+  private Calendar CalendarEnd;
+  private boolean isAllDay;
 
-  public Task(String title, String description, Calendar calendarStart) {
+  public Plan(String title, String description, Calendar CalendarStart, Calendar CalendarEnd, boolean isAllDay) {
     this.title = title;
     this.description = description;
-    CalendarStart = calendarStart;
+    this.CalendarStart = CalendarStart;
+    this.CalendarEnd = CalendarEnd;
+    this.isAllDay = isAllDay;
   }
 
   public String getTitle() {
@@ -35,5 +39,21 @@ public class Task {
 
   public void setCalendarStart(Calendar calendarStart) {
     CalendarStart = calendarStart;
+  }
+
+  public Calendar getCalendarEnd() {
+    return CalendarEnd;
+  }
+
+  public void setCalendarEnd(Calendar calendarEnd) {
+    CalendarEnd = calendarEnd;
+  }
+
+  public boolean isAllDay() {
+    return isAllDay;
+  }
+
+  public void setAllDay(boolean allDay) {
+    isAllDay = allDay;
   }
 }
