+++ app/src/main/java/com/example/mytodo/data/model/Task.java
@@ -1,20 +1,16 @@
-package com.example.mytodo;
+package com.example.mytodo.data.model;
 
 import java.util.Calendar;
 
-public class Plan {
+public class Task {
   private String title;
   private String description;
   private Calendar CalendarStart;
-  private Calendar CalendarEnd;
-  private boolean isAllDay;
 
-  public Plan(String title, String description, Calendar CalendarStart, Calendar CalendarEnd, boolean isAllDay) {
+  public Task(String title, String description, Calendar calendarStart) {
     this.title = title;
     this.description = description;
-    this.CalendarStart = CalendarStart;
-    this.CalendarEnd = CalendarEnd;
-    this.isAllDay = isAllDay;
+    CalendarStart = calendarStart;
   }
 
   public String getTitle() {
@@ -39,21 +35,5 @@ public class Plan {
 
   public void setCalendarStart(Calendar calendarStart) {
     CalendarStart = calendarStart;
-  }
-
-  public Calendar getCalendarEnd() {
-    return CalendarEnd;
-  }
-
-  public void setCalendarEnd(Calendar calendarEnd) {
-    CalendarEnd = calendarEnd;
-  }
-
-  public boolean isAllDay() {
-    return isAllDay;
-  }
-
-  public void setAllDay(boolean allDay) {
-    isAllDay = allDay;
   }
 }
