+++ app/src/main/java/com/example/mytodo/ui/main/DayFragment.java
@@ -1,10 +1,9 @@
-package com.example.mytodo;
+package com.example.mytodo.ui.main;
 
 import android.content.Context;
 import android.graphics.Rect;
 import android.os.Bundle;
 import android.os.Handler;
-import android.util.Log;
 import android.view.GestureDetector;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -14,10 +13,8 @@ import android.widget.ScrollView;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import com.example.mytodo.MainActivity.GestureListener;
-import com.google.android.material.tabs.TabLayout;
+import com.example.mytodo.R;
 import java.util.Calendar;
-import java.util.Date;
 
 public class DayFragment extends Fragment {
   private MainActivity mainActivity;
