+++ app/src/main/java/com/example/mytodo/ui/main/WeekFragment.java
@@ -1,4 +1,4 @@
-package com.example.mytodo;
+package com.example.mytodo.ui.main;
 
 import android.content.Context;
 import android.graphics.Rect;
@@ -13,6 +13,7 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import com.example.mytodo.R;
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
