+++ app/src/main/res/values/styles.xml
@@ -134,7 +134,7 @@
     <item name="android:layout_weight">1</item>
   </style>
 
-  <!--AddToDoActivity-->
+  <!--AddOrEditToDoActivity-->
   <style name="CustomPickerDialog" parent="Theme.AppCompat.Dialog.Alert">
     <!-- ダイアログのサイズ -->
     <item name="android:windowMinWidthMajor">20%</item>
